Splitting the code to make the blogAdmin smaller
